+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">Maquiagem</string>
+    <string name="edit_type">Tipo</string>
 </resources>
